Move GUI adapter into package and update imports
