+++ kielproc_monorepo/kielproc/gui_adapter.py
